refactor(accordion): decouple AccordionItemContext for reusability
